neurons: SpikingNeuron class, doesn't yet consider weights
Just implemented a basic SpikingNeuron class.  The activations and
firings are implemented using Events on private and public EventBuses.

Each neuron has a firing bus on to which it fires broadcasts of its
firings.  A network will have its own EventBus to update weights etc.
As of now, the weights are not considered inside the neuron.  This is
because a network class hasn't been implemented yet.

Soon, very soon!